Add files via upload
Its a Brick Breaker Game which is play by a ball to destroy the bricks and use paddle to throw a ball in left or right directions.
It is implemented in java by using map concept and use swing features such as frame, panel, button etc. and AWT features such as  graphics, timer, key listener and ActionListener.